update package structure to make intellij idea happy
