+++ forsuredbapi/src/main/java/com/fsryan/forsuredb/api/FSDocStoreGetApi.java
@@ -61,7 +61,7 @@ public interface FSDocStoreGetApi<T> extends FSGetApi {
      * </p>
      * <p>
      *     Note that this will most-likely lead to errors if you change the fields of this class and their
-     *     serialization
+     *     jacksonserialization
      * </p>
      * @param retriever a {@link Retriever} which points to a set of results for this {@link FSDocStoreGetApi}
      * @return an object of type S deseralized from the string in the doc column
@@ -71,7 +71,7 @@ public interface FSDocStoreGetApi<T> extends FSGetApi {
 
     /**
      * <p>
-     *     Use this only if you want to perform your own deserialization and that the initial serialization
+     *     Use this only if you want to perform your own deserialization and that the initial jacksonserialization
      *     was to a String.
      * </p>
      * @param retriever a {@link Retriever} which points to a set of results for this {@link FSDocStoreGetApi}
@@ -81,7 +81,7 @@ public interface FSDocStoreGetApi<T> extends FSGetApi {
 
     /**
      * <p>
-     *     Use this only if you want to perform your own deserialization and that the original serialization
+     *     Use this only if you want to perform your own deserialization and that the original jacksonserialization
      *     was into a byte array.
      * </p>
      * @param retriever a {@link Retriever} which points to a set of results for this {@link FSDocStoreGetApi}
@@ -93,7 +93,7 @@ public interface FSDocStoreGetApi<T> extends FSGetApi {
     /**
      * <p>
      *     Meta-data for the record that allows you to select all records by their java class. This
-     *     column has nothing to do with serialization/deserialization. And thus, if you refactor
+     *     column has nothing to do with jacksonserialization/deserialization. And thus, if you refactor
      *     the package name or the class name for a class, then this will not be reflected in this
      *     column . . . yet.
      * </p>
